NutriScan_Deploy/requirements.txt
streamlit
transformers
pillow
numpy
torch
torchvision
pytz
plotly